Extract Event things to own file, leaving turbomodule file cleaner
